Add one-r technique as a strategy for building a model
One-r looks at all the values of a column and builds a rule based on the
average sale price of all the houses with that value.  This was done for
all columns in the dataset and then tested against the training data to
see which columns produced the most correct results.  After throwing out
the most overfit columns, OverallQual was used to build a rule to run
against the test data.

The test function is currently being built in the runner program and
uses some hard coded values in that process.  This should be cleaned up,
but it felt like the work so far should be committed.