knn test cleanup
Changes to the generation of the test files were made so that the
average of the nearest neighbors is a float.  Changed the expected
values in the tests to also be floats

Also added a parameter to the function that generates the test function
so that it can have a value for K passed in rather than using the
default value that is defined in the code.