Create new knn distance function
Created a new distance function for the knn algorithm.  A multimethod
based on the column being compared chooses between a custom function for
that column which looks at the numeric difference between two values and
returns a normalized proportion of the difference and the maximum
difference of all values for that column.